requirements.txt
beautifulsoup4
pandas
sqlalchemy
luigi
playwright
python-dotenv
psycopg2-binary
